Fix edge case bug for explode()
